Let api-promise be composed of api-rx
